[TASK] Refactored Hooks directory
